Atualização e implementação de novas funcionalidades (#2)
* Criar o dockerfile - versão inicial

* Criar o dockerfile - versão 20.04 - teste 1

* Criar o dockerfile - versão 20.04 - teste 2

* Criar o dockerfile - versão 20.04 - ajustar caminho do composer

* Colocar dockerfile na raiz

* Manter locale padrão

* clear typo

* adicionar dumpautoload

* iniciar rota de monitoramento

* corrigir supervisor

* criar entrypoint

* adiciona vim ao ambiente para ediçao de arquivo

* adiciona arquivo de deployment com variáveis de ambiente

* teste de variávies no jenkins

* adiciona iproute2

* corrigir caminho das pasta para permissões

* corrigir permissão da pasta storage

* corrigir permissão da pasta storage

* Inicia o Cron

* Ajuste na execução do cron para buscar todos os dados de histórico, cronograma e empenhos

* Insere telnet no container

* logo cgu

* Ajustar execução do script de importação

* Atualizar imagens

* ignorar data

* Remover arquivos não utilizados

* Atualizar readme inicial

* Ajustar home

* Ajustar home

* Iniciar redução do container da imagem principal

* Atualizar versão do laravel e abstrair aplicação do framework

* Prepostos

* Adicionar Faturas dos contratos

* Adicionar Responsáveis

* Adicionar id original de histórico

* Adicona id original para cronograma

* Adicionar tabela de arquivos

* Atualizar versão do php para 8.1

* Adicionar pasta php 8.1

* Adicionar feedback na importação dos dados via email

* Adiciona Action para fatura

* Alterar funções para actions

* Adicionar opção --all para importação dos contratos

* clean up

* Atualizar documentação

* Ajusta agendamento

* Remover dados antes de inserir novo - não manter dados na base local que não existem na base do comprasnet

* Adicionar tipo de retorno da action

* Adicionar IdEmpenhoOriginal

* Adicionar fatura empenho e exclusão de faturas e empenhos

* Action para adicionar fatura empenho

* Alterar id para biginteger

* Adicionar Fatura Mes e Ano

* Alterar nulidade do empenho em caso de alteração da resposta da api

* Adicionar itens vinculados à fatura

* Reduzir nomes das colunas

* Adicionar publucações

* Padronizar nomes das tabelas e campos conforme roteiro CGU

* Atualiar rotas diretas da API com timeout

* Itens de contratos

* Ajustar o item do contrato

* Atualiza tipo de dado

* Remove colunas de links

* Criar ação comum para duplicidae de resposta na API

* Corrigir mensagem de log

* Migrar tabelas no depoy

* Atualizar comando de importação geral do sistema

* Atualizar documentação

* Corrigir link

* Atualizar documentação novos comandos

* Ajustar documentação

* Criar monitoramento

* Ajustar sql

* Ajustar imagens

* readme docs

* Atualizar documentação

* Ajustar validação de array para retornar qualquer tipo de dado

* Ajustar tabelas com novos campos

* Log no output do container para saber o resultado do monitoramento

* Ajustar monitoramento

* Alterar quantidade para decimal - erro API

* Não colocar erro no log

* Remover debug

* Adicionar opção de importar depois de determinado contrato

* Configurar log para erro na importação de contratos

* Criar log de atividades do sistema

* Padronizar saídas de erro da aplicação

* Ajustar configuração de email

* Alterar tipo de daddo na coluna detalhamento execução em log

* Ajustar configuração do smtp

+++ database/migrations/2020_01_30_193556_create_table_contrato.php
@@ -24,7 +24,7 @@ class CreateTableContrato extends Migration
         });
 
         $schema->create('Contrato', function (Blueprint $table) {
-            $table->integer('IdContrato');
+            $table->bigInteger('IdContrato');
             $table->varChar('TxtReceitaDespesa')->nullable();
             $table->varChar('NumContrato')->nullable();
